requirements.txt
flask            # For building the backend server
python-docx      # For processing .docx files
